editor: keep the light position in double through the panel state
At large world coordinates (~1e5) LightState.pos was a float[] holding the
ABSOLUTE position: the gizmo's per-frame write-back re-snapped the light to
the ~8mm float lattice every frame of a drag (stuttering, creeping handle),
and LightSync.push wrote the quantized value back into the double
PlacedLight. pos is double[] now; the gizmo write-back and place-at-eye
stay in double; Widgets gains a double dragValue variant for the X/Y/Z
fields (drag delta applied to the double start value). The gizmo model was
already camera-relative - only the storage narrowed.

+++ src/client/java/org/qualet/irlredactor/editor/LightSync.java
@@ -63,7 +63,7 @@ public final class LightSync
         s.cookieScale[0] = l.cookieScale;
         s.cookieInvert.set(l.cookieInvert);
 
-        s.pos[0] = (float) l.x; s.pos[1] = (float) l.y; s.pos[2] = (float) l.z;
+        s.pos[0] = l.x; s.pos[1] = l.y; s.pos[2] = l.z;
         s.dir[0] = l.dirX; s.dir[1] = l.dirY; s.dir[2] = l.dirZ;
     }
